backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
requests==2.31.0

# Text extraction dependencies
PyPDF2==3.0.1
python-docx==1.1.0
textract==1.6.5
chardet==3.0.4

# Database dependencies
sqlalchemy==2.0.23

# Text similarity and NLP
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
scikit-learn==1.3.2

# Additional utilities
regex==2023.10.3
unicodedata2==15.1.0 